+++ src/SampleCode.csproj
@@ -93,26 +93,26 @@
     <Compile Include="Samples\Authentication\StandAloneHttpSignature.cs" />
     <Compile Include="Samples\Authentication\StandAloneJWT.cs" />
     <Compile Include="Samples\Reporting\CoreServices\CreateStandardOrClassicSubscription.cs" />
-    <Compile Include="Samples\Risk Management\CoreServices\ListManagement.cs" />
-    <Compile Include="Samples\Risk Management\CoreServices\AddDuplicateInformation.cs" />
-    <Compile Include="Samples\Risk Management\CoreServices\AddressMatchNotFound.cs" />
-    <Compile Include="Samples\Risk Management\CoreServices\ApartmentNumberMissingOrNotFound.cs" />
-    <Compile Include="Samples\Risk Management\CoreServices\CanadianBillingDetails.cs" />
-    <Compile Include="Samples\Risk Management\CoreServices\ComplianceStatusCompleted.cs" />
-    <Compile Include="Samples\Risk Management\CoreServices\CreateDecisionManagerCase.cs" />
-    <Compile Include="Samples\Risk Management\CoreServices\DecisionProfileReject.cs" />
-    <Compile Include="Samples\Risk Management\CoreServices\DmWithBuyerInformation.cs" />
-    <Compile Include="Samples\Risk Management\CoreServices\DmWithDeviceInformation.cs" />
-    <Compile Include="Samples\Risk Management\CoreServices\DmWithMerchantDefinedInformation.cs" />
-    <Compile Include="Samples\Risk Management\CoreServices\DmWithShippingInformation.cs" />
-    <Compile Include="Samples\Risk Management\CoreServices\DmWithTravelInformation.cs" />
-    <Compile Include="Samples\Risk Management\CoreServices\ExportComplianceInformationProvided.cs" />
-    <Compile Include="Samples\Risk Management\CoreServices\FraudMarking.cs" />
-    <Compile Include="Samples\Risk Management\CoreServices\MultipleLineItems.cs" />
-    <Compile Include="Samples\Risk Management\CoreServices\MultipleSanctionLists.cs" />
-    <Compile Include="Samples\Risk Management\CoreServices\NoCompanyName.cs" />
-    <Compile Include="Samples\Risk Management\CoreServices\RemoveFromHistory.cs" />
-    <Compile Include="Samples\Risk Management\CoreServices\ScoreExceedsThreshold.cs" />
+    <Compile Include="Samples\RiskManagement\CoreServices\ListManagement.cs" />
+    <Compile Include="Samples\RiskManagement\CoreServices\AddDuplicateInformation.cs" />
+    <Compile Include="Samples\RiskManagement\CoreServices\AddressMatchNotFound.cs" />
+    <Compile Include="Samples\RiskManagement\CoreServices\ApartmentNumberMissingOrNotFound.cs" />
+    <Compile Include="Samples\RiskManagement\CoreServices\CanadianBillingDetails.cs" />
+    <Compile Include="Samples\RiskManagement\CoreServices\ComplianceStatusCompleted.cs" />
+    <Compile Include="Samples\RiskManagement\CoreServices\CreateDecisionManagerCase.cs" />
+    <Compile Include="Samples\RiskManagement\CoreServices\DecisionProfileReject.cs" />
+    <Compile Include="Samples\RiskManagement\CoreServices\DmWithBuyerInformation.cs" />
+    <Compile Include="Samples\RiskManagement\CoreServices\DmWithDeviceInformation.cs" />
+    <Compile Include="Samples\RiskManagement\CoreServices\DmWithMerchantDefinedInformation.cs" />
+    <Compile Include="Samples\RiskManagement\CoreServices\DmWithShippingInformation.cs" />
+    <Compile Include="Samples\RiskManagement\CoreServices\DmWithTravelInformation.cs" />
+    <Compile Include="Samples\RiskManagement\CoreServices\ExportComplianceInformationProvided.cs" />
+    <Compile Include="Samples\RiskManagement\CoreServices\FraudMarking.cs" />
+    <Compile Include="Samples\RiskManagement\CoreServices\MultipleLineItems.cs" />
+    <Compile Include="Samples\RiskManagement\CoreServices\MultipleSanctionLists.cs" />
+    <Compile Include="Samples\RiskManagement\CoreServices\NoCompanyName.cs" />
+    <Compile Include="Samples\RiskManagement\CoreServices\RemoveFromHistory.cs" />
+    <Compile Include="Samples\RiskManagement\CoreServices\ScoreExceedsThreshold.cs" />
     <Compile Include="Samples\Flex\CoreServices\GenerateKey.cs" />
     <Compile Include="Configuration.cs" />
     <Compile Include="Samples\Flex\CoreServices\TokenizeCard.cs" />
@@ -158,9 +158,9 @@
     <Compile Include="Samples\Reporting\CoreServices\GetReportingResourceInformation.cs" />
     <Compile Include="Samples\Reporting\CoreServices\GetSubscriptionForReportName.cs" />
     <Compile Include="Samples\Reporting\CoreServices\RetrieveAvailableReports.cs" />
-    <Compile Include="Samples\Risk Management\CoreServices\ShippingDetailsNotUSOrCanada.cs" />
-    <Compile Include="Samples\Risk Management\CoreServices\ValidateExportCompliance.cs" />
-    <Compile Include="Samples\Risk Management\CoreServices\VerifyCustomerAddress.cs" />
+    <Compile Include="Samples\RiskManagement\CoreServices\ShippingDetailsNotUSOrCanada.cs" />
+    <Compile Include="Samples\RiskManagement\CoreServices\ValidateExportCompliance.cs" />
+    <Compile Include="Samples\RiskManagement\CoreServices\VerifyCustomerAddress.cs" />
     <Compile Include="Samples\SecureFileShare\CoreServices\DownloadFileWithFileIdentifier.cs" />
     <Compile Include="Samples\SecureFileShare\CoreServices\GetListOfFiles.cs" />
     <Compile Include="Samples\TMS\CoreServices\CreateInstrumentIdentifier.cs" />
