+++ src/Samples/RiskManagement/CoreServices/DmWithMerchantDefinedInformation.cs
@@ -6,9 +6,9 @@ using System.Threading.Tasks;
 using CyberSource.Api;
 using CyberSource.Model;
 
-namespace Cybersource_rest_samples_dotnet.Samples.Risk_Management.CoreServices
+namespace Cybersource_rest_samples_dotnet.Samples.RiskManagement.CoreServices
 {
-    public class DmWithTravelInformation
+    public class DmWithMerchantDefinedInformation
     {
         public static RiskV1DecisionsPost201Response Run()
         {
@@ -53,26 +53,19 @@ namespace Cybersource_rest_samples_dotnet.Samples.Risk_Management.CoreServices
 
             requestObj.OrderInformation = orderInformation;
 
-            var travelInformation = new Riskv1decisionsTravelInformation();
+            var merchantDefinedInformation = new List<Riskv1decisionsMerchantDefinedInformation>();
 
-            travelInformation.CompleteRoute = "SFO-JFK:JFK-BLR";
-            travelInformation.DepartureTime = "2011-03-20 11:30pm GMT";
-            travelInformation.JourneyType = "One way";
-            var legs = new List<Riskv1decisionsTravelInformationLegs>();
+            var merchantDefinedInformation0 = new Riskv1decisionsMerchantDefinedInformation();
+            merchantDefinedInformation0.Key = "1";
+            merchantDefinedInformation0.Value = "Test";
+            merchantDefinedInformation.Add(merchantDefinedInformation0);
 
-            var legs0 = new Riskv1decisionsTravelInformationLegs();
-            legs0.Origination = "SFO";
-            legs0.Destination = "JFK";
-            legs.Add(legs0);
+            var merchantDefinedInformation1 = new Riskv1decisionsMerchantDefinedInformation();
+            merchantDefinedInformation1.Key = "2";
+            merchantDefinedInformation1.Value = "Test2";
+            merchantDefinedInformation.Add(merchantDefinedInformation1);
 
-            var legs1 = new Riskv1decisionsTravelInformationLegs();
-            legs1.Origination = "JFK";
-            legs1.Destination = "BLR";
-            legs.Add(legs1);
-
-            travelInformation.Legs = legs;
-
-            requestObj.TravelInformation = travelInformation;
+            requestObj.MerchantDefinedInformation = merchantDefinedInformation;
 
             try
             {
